% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluralize_gift.R
\name{pluralize_gift}
\alias{pluralize_gift}
\title{Takes a noun and makes it plural}
\usage{
pluralize_gift(gift)
}
\arguments{
\item{gift}{A string or vector of strings}
}
\value{
A string or vector of strings with the pluralized words
}
\description{
Takes a noun and makes it plural
}
