% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_phrase.R
\name{make_phrase}
\alias{make_phrase}
\title{Takes a noun and makes it plural}
\usage{
make_phrase(num, num_word, item, verb, adjective, location)
}
\arguments{
\item{num}{An integer}

\item{num_word}{A string corresponding to the integer}

\item{item}{A string}

\item{verb}{A string}

\item{adjective}{A string}

\item{location}{A string}
}
\value{
A string containing the words in grammatical order.
}
\description{
Takes a noun and makes it plural
}
