% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sing_all_untill_day.R
\name{sing_all_untill_day}
\alias{sing_all_untill_day}
\title{Produces the string for one day of the song.}
\usage{
sing_all_untill_day(dataset, max_day, phrase_col)
}
\arguments{
\item{dataset}{A data frame containing information about gifts}

\item{max_day}{The number of the max_day for the day you want to sing up to}

\item{phrase_col}{The variable name for the column in the dataset that
contains the gift phrases}
}
\value{
A string singing the max_day of the song with all gifts for the given day.
}
\description{
Produces the string for one day of the song.
}
